# Do not add nulls to the hash table when null_equality::NOT_EQUAL is passed to left_semi_join and left_anti_join