--- a/cpp/benchmarks/join/join_benchmark.cu
+++ b/cpp/benchmarks/join/join_benchmark.cu
@@ -37,8 +37,8 @@ template <typename key_type, typename payload_type>
 class Join : public cudf::benchmark {
 };
 
-template <typename key_type, typename payload_type, bool Nullable>
-static void BM_join(benchmark::State &state)
+template <typename key_type, typename payload_type, bool Nullable, typename Join>
+static void BM_join(Join JoinFunc, benchmark::State& state)
 {
   const cudf::size_type build_table_size{(cudf::size_type)state.range(0)};
   const cudf::size_type probe_table_size{(cudf::size_type)state.range(1)};
@@ -105,20 +105,69 @@ static void BM_join(benchmark::State &state)
   for (auto _ : state) {
     cuda_event_timer raii(state, true, rmm::cuda_stream_default);
 
-    auto result = cudf::inner_join(
+    auto result = JoinFunc(
       probe_table, build_table, columns_to_join, columns_to_join, cudf::null_equality::UNEQUAL);
   }
 }
 
-#define JOIN_BENCHMARK_DEFINE(name, key_type, payload_type, nullable) \
-  BENCHMARK_TEMPLATE_DEFINE_F(Join, name, key_type, payload_type)     \
-  (::benchmark::State & st) { BM_join<key_type, payload_type, nullable>(st); }
+#define JOIN_BENCHMARK_DEFINE(name, key_type, payload_type, nullable)         \
+  BENCHMARK_TEMPLATE_DEFINE_F(Join, name, key_type, payload_type)             \
+  (::benchmark::State & st)                                                   \
+  {                                                                           \
+    auto join = [](cudf::table_view const& left,                              \
+                   cudf::table_view const& right,                             \
+                   std::vector<cudf::size_type> const& left_on,               \
+                   std::vector<cudf::size_type> const& right_on,              \
+                   cudf::null_equality compare_nulls) {                       \
+      return cudf::inner_join(left, right, left_on, right_on, compare_nulls); \
+    };                                                                        \
+    BM_join<key_type, payload_type, nullable>(join, st);                      \
+  }
 
 JOIN_BENCHMARK_DEFINE(join_32bit, int32_t, int32_t, false);
 JOIN_BENCHMARK_DEFINE(join_64bit, int64_t, int64_t, false);
 JOIN_BENCHMARK_DEFINE(join_32bit_nulls, int32_t, int32_t, true);
 JOIN_BENCHMARK_DEFINE(join_64bit_nulls, int64_t, int64_t, true);
 
+#define LEFT_ANTI_JOIN_BENCHMARK_DEFINE(name, key_type, payload_type, nullable)   \
+  BENCHMARK_TEMPLATE_DEFINE_F(Join, name, key_type, payload_type)                 \
+  (::benchmark::State & st)                                                       \
+  {                                                                               \
+    auto join = [](cudf::table_view const& left,                                  \
+                   cudf::table_view const& right,                                 \
+                   std::vector<cudf::size_type> const& left_on,                   \
+                   std::vector<cudf::size_type> const& right_on,                  \
+                   cudf::null_equality compare_nulls) {                           \
+      return cudf::left_anti_join(left, right, left_on, right_on, compare_nulls); \
+    };                                                                            \
+    BM_join<key_type, payload_type, nullable>(join, st);                          \
+  }
+
+LEFT_ANTI_JOIN_BENCHMARK_DEFINE(left_anti_join_32bit, int32_t, int32_t, false);
+LEFT_ANTI_JOIN_BENCHMARK_DEFINE(left_anti_join_64bit, int64_t, int64_t, false);
+LEFT_ANTI_JOIN_BENCHMARK_DEFINE(left_anti_join_32bit_nulls, int32_t, int32_t, true);
+LEFT_ANTI_JOIN_BENCHMARK_DEFINE(left_anti_join_64bit_nulls, int64_t, int64_t, true);
+
+#define LEFT_SEMI_JOIN_BENCHMARK_DEFINE(name, key_type, payload_type, nullable)   \
+  BENCHMARK_TEMPLATE_DEFINE_F(Join, name, key_type, payload_type)                 \
+  (::benchmark::State & st)                                                       \
+  {                                                                               \
+    auto join = [](cudf::table_view const& left,                                  \
+                   cudf::table_view const& right,                                 \
+                   std::vector<cudf::size_type> const& left_on,                   \
+                   std::vector<cudf::size_type> const& right_on,                  \
+                   cudf::null_equality compare_nulls) {                           \
+      return cudf::left_semi_join(left, right, left_on, right_on, compare_nulls); \
+    };                                                                            \
+    BM_join<key_type, payload_type, nullable>(join, st);                          \
+  }
+
+LEFT_SEMI_JOIN_BENCHMARK_DEFINE(left_semi_join_32bit, int32_t, int32_t, false);
+LEFT_SEMI_JOIN_BENCHMARK_DEFINE(left_semi_join_64bit, int64_t, int64_t, false);
+LEFT_SEMI_JOIN_BENCHMARK_DEFINE(left_semi_join_32bit_nulls, int32_t, int32_t, true);
+LEFT_SEMI_JOIN_BENCHMARK_DEFINE(left_semi_join_64bit_nulls, int64_t, int64_t, true);
+
+// join -----------------------------------------------------------------------
 BENCHMARK_REGISTER_F(Join, join_32bit)
   ->Unit(benchmark::kMillisecond)
   ->Args({100'000, 100'000})
@@ -144,12 +193,86 @@ BENCHMARK_REGISTER_F(Join, join_32bit_nulls)
   ->Args({100'000, 1'000'000})
   ->Args({10'000'000, 10'000'000})
   ->Args({10'000'000, 40'000'000})
-  ->Args({10'000'000, 100'000'000})
+  ->Args({10'000'00, 100'000'00})
   ->Args({100'000'000, 100'000'000})
   ->Args({80'000'000, 240'000'000})
   ->UseManualTime();
 
 BENCHMARK_REGISTER_F(Join, join_64bit_nulls)
+  ->Unit(benchmark::kMillisecond)
+  ->Args({50'000'000, 50'000'000})
+  ->Args({40'000'000, 120'000'000})
+  ->UseManualTime();
+
+// left anti-join -------------------------------------------------------------
+BENCHMARK_REGISTER_F(Join, left_anti_join_32bit)
+  ->Unit(benchmark::kMillisecond)
+  ->Args({100'000, 100'000})
+  ->Args({100'000, 400'000})
+  ->Args({100'000, 1'000'000})
+  ->Args({10'000'000, 10'000'000})
+  ->Args({10'000'000, 40'000'000})
+  ->Args({10'000'000, 100'000'000})
+  ->Args({100'000'000, 100'000'000})
+  ->Args({80'000'000, 240'000'000})
+  ->UseManualTime();
+
+BENCHMARK_REGISTER_F(Join, left_anti_join_64bit)
+  ->Unit(benchmark::kMillisecond)
+  ->Args({50'000'000, 50'000'000})
+  ->Args({40'000'000, 120'000'000})
+  ->UseManualTime();
+
+BENCHMARK_REGISTER_F(Join, left_anti_join_32bit_nulls)
+  ->Unit(benchmark::kMillisecond)
+  ->Args({100'000, 100'000})
+  ->Args({100'000, 400'000})
+  ->Args({100'000, 1'000'000})
+  ->Args({10'000'000, 10'000'000})
+  ->Args({10'000'000, 40'000'000})
+  ->Args({10'000'00, 100'000'00})
+  ->Args({100'000'000, 100'000'000})
+  ->Args({80'000'000, 240'000'000})
+  ->UseManualTime();
+
+BENCHMARK_REGISTER_F(Join, left_anti_join_64bit_nulls)
+  ->Unit(benchmark::kMillisecond)
+  ->Args({50'000'000, 50'000'000})
+  ->Args({40'000'000, 120'000'000})
+  ->UseManualTime();
+
+// left semi-join -------------------------------------------------------------
+BENCHMARK_REGISTER_F(Join, left_semi_join_32bit)
+  ->Unit(benchmark::kMillisecond)
+  ->Args({100'000, 100'000})
+  ->Args({100'000, 400'000})
+  ->Args({100'000, 1'000'000})
+  ->Args({10'000'000, 10'000'000})
+  ->Args({10'000'000, 40'000'000})
+  ->Args({10'000'000, 100'000'000})
+  ->Args({100'000'000, 100'000'000})
+  ->Args({80'000'000, 240'000'000})
+  ->UseManualTime();
+
+BENCHMARK_REGISTER_F(Join, left_semi_join_64bit)
+  ->Unit(benchmark::kMillisecond)
+  ->Args({50'000'000, 50'000'000})
+  ->Args({40'000'000, 120'000'000})
+  ->UseManualTime();
+
+BENCHMARK_REGISTER_F(Join, left_semi_join_32bit_nulls)
+  ->Unit(benchmark::kMillisecond)
+  ->Args({100'000, 100'000})
+  ->Args({100'000, 400'000})
+  ->Args({100'000, 1'000'000})
+  ->Args({10'000'000, 10'000'000})
+  ->Args({10'000'000, 40'000'000})
+  ->Args({10'000'00, 100'000'00})
+  ->Args({100'000'000, 100'000'000})
+  ->Args({80'000'000, 240'000'000})
+  ->UseManualTime();
+
+BENCHMARK_REGISTER_F(Join, left_semi_join_64bit_nulls)
   ->Unit(benchmark::kMillisecond)
   ->Args({50'000'000, 50'000'000})
   ->Args({40'000'000, 120'000'000})