--- a/cpp/src/join/semi_join.cu
+++ b/cpp/src/join/semi_join.cu
@@ -92,7 +92,7 @@ std::unique_ptr<rmm::device_uvector<cudf::size_type>> left_semi_anti_join(
                                                 equality_build);
   auto hash_table     = *hash_table_ptr;
 
-  // if compare_nulls == NOT_EQUAL, we can simply ignore any rows that are
+  // if compare_nulls == UNEQUAL, we can simply ignore any rows that
   // contain a NULL in any column as they will never compare to equal.
   auto const row_bitmask = (compare_nulls == null_equality::EQUAL)
                              ? rmm::device_buffer{}