… cpp/tests/join/semi_join_tests.cpp