@nvdbaranec 2021-05-18
I noticed as I was doing this that we have no tests for left_anti_join.  Will add.